React to API changes in aspnet-contrib/AspNet.Security.OpenIdConnect.Server
https://github.com/aspnet-contrib/AspNet.Security.OpenIdConnect.Server/commit/0fecdbe58ffff2e3fef2078ccc95196af9753662
