+++ src/OpenIddict.Core/OpenIddictProvider.Authentication.cs
@@ -1,0 +1,172 @@
+﻿/*
+ * Licensed under the Apache License, Version 2.0 (http://www.apache.org/licenses/LICENSE-2.0)
+ * See https://github.com/openiddict/core for more information concerning
+ * the license and the contributors participating to this project.
+ */
+
+using System;
+using System.Diagnostics;
+using System.IdentityModel.Tokens.Jwt;
+using System.Linq;
+using System.Security.Claims;
+using System.Threading.Tasks;
+using AspNet.Security.OpenIdConnect.Extensions;
+using AspNet.Security.OpenIdConnect.Server;
+using Microsoft.AspNet.Authentication;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Internal;
+
+namespace OpenIddict {
+    public partial class OpenIddictProvider<TUser, TApplication> : OpenIdConnectServerProvider where TUser : class where TApplication : class {
+        public override async Task ValidateAuthorizationRequest([NotNull] ValidateAuthorizationRequestContext context) {
+            var manager = context.HttpContext.RequestServices.GetRequiredService<OpenIddictManager<TUser, TApplication>>();
+
+            // Note: redirect_uri is not required for pure OAuth2 requests
+            // but this provider uses a stricter policy making it mandatory,
+            // as required by the OpenID Connect core specification.
+            // See http://openid.net/specs/openid-connect-core-1_0.html#AuthRequest.
+            if (string.IsNullOrEmpty(context.RedirectUri)) {
+                context.Reject(
+                    error: OpenIdConnectConstants.Errors.InvalidRequest,
+                    description: "The required redirect_uri parameter was missing.");
+
+                return;
+            }
+
+            // Retrieve the application details corresponding to the requested client_id.
+            var application = await manager.FindApplicationByIdAsync(context.ClientId);
+            if (application == null) {
+                context.Reject(
+                    error: OpenIdConnectConstants.Errors.InvalidClient,
+                    description: "Application not found in the database: ensure that your client_id is correct.");
+
+                return;
+            }
+
+            if (!await manager.ValidateRedirectUriAsync(application, context.RedirectUri)) {
+                context.Reject(
+                    error: OpenIdConnectConstants.Errors.InvalidClient,
+                    description: "Invalid redirect_uri.");
+
+                return;
+            }
+
+            // To prevent downgrade attacks, ensure that authorization requests using the hybrid/implicit
+            // flow are rejected if the client identifier corresponds to a confidential application.
+            // Note: when using the authorization code grant, ValidateClientAuthentication is responsible of
+            // rejecting the token request if the client_id corresponds to an unauthenticated confidential client.
+            if (await manager.IsConfidentialApplicationAsync(application) && !context.Request.IsAuthorizationCodeFlow()) {
+                context.Reject(
+                    error: OpenIdConnectConstants.Errors.InvalidRequest,
+                    description: "Confidential clients can only use response_type=code.");
+
+                return;
+            }
+
+            // Run additional checks for prompt=none requests.
+            if (string.Equals(context.Request.Prompt, "none", StringComparison.Ordinal)) {
+                // If the user is not authenticated, return an error to the client application.
+                // See http://openid.net/specs/openid-connect-core-1_0.html#Authenticates
+                if (!context.HttpContext.User.Identities.Any(identity => identity.IsAuthenticated)) {
+                    context.Reject(
+                        error: OpenIdConnectConstants.Errors.LoginRequired,
+                        description: "The user must be authenticated.");
+
+                    return;
+                }
+
+                // Extract the principal contained in the id_token_hint parameter.
+                // If no principal can be extracted, an error is returned to the client application.
+                var principal = await context.HttpContext.Authentication.AuthenticateAsync(context.Options.AuthenticationScheme);
+                if (principal == null) {
+                    context.Reject(
+                        error: OpenIdConnectConstants.Errors.InvalidRequest,
+                        description: "The required id_token_hint parameter is missing.");
+
+                    return;
+                }
+
+                // Ensure the client application is listed as a valid audience in the identity token
+                // and that the identity token corresponds to the authenticated user.
+                if (!principal.HasClaim(JwtRegisteredClaimNames.Aud, context.Request.ClientId) ||
+                    !principal.HasClaim(ClaimTypes.NameIdentifier, context.HttpContext.User.GetClaim(ClaimTypes.NameIdentifier))) {
+                    context.Reject(
+                        error: OpenIdConnectConstants.Errors.InvalidRequest,
+                        description: "The id_token_hint parameter is invalid.");
+
+                    return;
+                }
+
+                // Ensure the user profile still exists in the database.
+                var user = await manager.FindByIdAsync(principal.GetUserId());
+                if (user == null) {
+                    context.Reject(
+                        error: OpenIdConnectConstants.Errors.InvalidRequest,
+                        description: "The id_token_hint parameter is invalid.");
+
+                    return;
+                }
+
+                // Return an error if the username corresponds to the registered
+                // email address and if the "email" scope has not been requested.
+                if (context.Request.HasScope(OpenIdConnectConstants.Scopes.Profile) &&
+                   !context.Request.HasScope(OpenIdConnectConstants.Scopes.Email) &&
+                    string.Equals(await manager.GetUserNameAsync(user),
+                                  await manager.GetEmailAsync(user),
+                                  StringComparison.OrdinalIgnoreCase)) {
+                    context.Reject(
+                        error: OpenIdConnectConstants.Errors.InvalidRequest,
+                        description: "The 'email' scope is required.");
+
+                    return;
+                }
+            }
+
+            context.Validate();
+        }
+
+        public override async Task AuthorizationEndpoint([NotNull] AuthorizationEndpointContext context) {
+            // Only handle prompt=none requests at this stage.
+            if (!string.Equals(context.Request.Prompt, "none", StringComparison.Ordinal)) {
+                return;
+            }
+
+            var manager = context.HttpContext.RequestServices.GetRequiredService<OpenIddictManager<TUser, TApplication>>();
+
+            // Note: principal is guaranteed to be non-null since ValidateAuthorizationRequest
+            // rejects prompt=none requests missing or having an invalid id_token_hint.
+            var principal = await context.HttpContext.Authentication.AuthenticateAsync(context.Options.AuthenticationScheme);
+            Debug.Assert(principal != null);
+
+            // Note: user may be null if the user was removed after
+            // the initial check made by ValidateAuthorizationRequest.
+            // In this case, ignore the prompt=none request and
+            // continue to the next middleware in the pipeline.
+            var user = await manager.FindByIdAsync(principal.GetUserId());
+            if (user == null) {
+                return;
+            }
+
+            // Note: filtering the username is not needed at this stage as OpenIddictController.Accept
+            // and OpenIddictProvider.GrantResourceOwnerCredentials are expected to reject requests that
+            // don't include the "email" scope if the username corresponds to the registed email address.
+            var identity = await manager.CreateIdentityAsync(user, context.Request.GetScopes());
+            Debug.Assert(identity != null);
+
+            // Create a new authentication ticket holding the user identity.
+            var ticket = new AuthenticationTicket(
+                new ClaimsPrincipal(identity),
+                null, context.Options.AuthenticationScheme);
+
+            ticket.SetResources(context.Request.GetResources());
+            ticket.SetScopes(context.Request.GetScopes());
+
+            // Call SignInAsync to create and return a new OpenID Connect response containing the serialized code/tokens.
+            await context.HttpContext.Authentication.SignInAsync(ticket.AuthenticationScheme, ticket.Principal, ticket.Properties);
+
+            // Mark the response as handled
+            // to skip the rest of the pipeline.
+            context.HandleResponse();
+        }
+    }
+}
