+++ src/OpenIddict.Core/OpenIddictProvider.Session.cs
@@ -1,0 +1,38 @@
+﻿/*
+ * Licensed under the Apache License, Version 2.0 (http://www.apache.org/licenses/LICENSE-2.0)
+ * See https://github.com/openiddict/core for more information concerning
+ * the license and the contributors participating to this project.
+ */
+
+using System.Threading.Tasks;
+using AspNet.Security.OpenIdConnect.Extensions;
+using AspNet.Security.OpenIdConnect.Server;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Internal;
+
+namespace OpenIddict {
+    public partial class OpenIddictProvider<TUser, TApplication> : OpenIdConnectServerProvider where TUser : class where TApplication : class {
+        public override async Task ValidateLogoutRequest([NotNull] ValidateLogoutRequestContext context) {
+            var manager = context.HttpContext.RequestServices.GetRequiredService<OpenIddictManager<TUser, TApplication>>();
+
+            // Skip validation if the optional post_logout_redirect_uri
+            // parameter was missing from the logout request.
+            if (string.IsNullOrEmpty(context.PostLogoutRedirectUri)) {
+                context.Skip();
+
+                return;
+            }
+
+            var application = await manager.FindApplicationByLogoutRedirectUri(context.PostLogoutRedirectUri);
+            if (application == null) {
+                context.Reject(
+                    error: OpenIdConnectConstants.Errors.InvalidClient,
+                    description: "Invalid post_logout_redirect_uri.");
+
+                return;
+            }
+
+            context.Validate();
+        }
+    }
+}
